services/user_service/requirements.txt
django
gunicorn
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
ldap3
drf-spectacular
drf-spectacular-sidecar
python-dotenv